# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Book a Seat – Workspace & Meeting Room Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /desks:
    get:
      operationId: listDesks
      tags: [desks]
      summary: List desks with optional availability filter
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: floor
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of desks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Desk"
    post:
      operationId: createDesk
      tags: [desks]
      summary: Create a new desk (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeskInput"
      responses:
        "201":
          description: Created desk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Desk"
  /desks/{id}:
    get:
      operationId: getDesk
      tags: [desks]
      summary: Get desk by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Desk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Desk"
    patch:
      operationId: updateDesk
      tags: [desks]
      summary: Update desk (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeskUpdate"
      responses:
        "200":
          description: Updated desk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Desk"
    delete:
      operationId: deleteDesk
      tags: [desks]
      summary: Delete desk (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /desks/{id}/maintenance:
    post:
      operationId: markDeskMaintenance
      tags: [desks]
      summary: Mark desk under maintenance (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "200":
          description: Updated desk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Desk"
  /desks/{id}/restore:
    post:
      operationId: restoreDesk
      tags: [desks]
      summary: Restore desk from maintenance (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated desk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Desk"

  # DESK BOOKINGS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Desk:
      type: object
      required: [id, name, floor, status, qrCode, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        floor:
          type: string
        zone:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, booked, maintenance]
        amenities:
          type: ["string", "null"]
        qrCode:
          type: string
        isAvailable:
          type: boolean
        createdAt:
          type: string
    DeskInput:
      type: object
      required: [name, floor]
      properties:
        name:
          type: string
        floor:
          type: string
        zone:
          type: string
        amenities:
          type: string
    DeskUpdate:
      type: object
      properties:
        name:
          type: string
        floor:
          type: string
        zone:
          type: string
        amenities:
          type: string
        status:
          type: string
    MaintenanceInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...